requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==18.2.0
backports.functools-lru-cache==1.3
beautifulsoup4==4.6.3
bleach==3.1.0
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Cython==0.29.5
DAWG-Python==0.7.2
decorator==4.3.2
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.3
idna==2.8
intervaltree==2.1.0
ipykernel==5.1.0
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
mistune==0.8.4
more-itertools==6.0.0
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pybind11==2.2.4
Pygments==2.3.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pytest==4.3.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
pyzmq==18.0.0
qtconsole==4.4.3
razdel==0.2.0
requests==2.21.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sortedcontainers==2.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
yargy==0.11.0
